+++ src/api.ts
@@ -1,0 +1,5 @@
+import { getTextLanguage } from "./utils/str";
+
+const utils = { getTextLanguage };
+
+export default { utils };